fix(test): consolidate _drain helper and document yield patterns
Consensus-driven changes from expert panel review:
- Consolidate _drain() to tests/conftest.py with TIMEOUT_IO default
- Add inline comments explaining consecutive yield_once() calls
- Remove duplicate _drain definitions from core/conftest and harness
